+++ test/test_socket_options.cpp
@@ -123,7 +123,7 @@ TEST_F(TestSocketOptions, LossMaxTTL)
     int opt_len = 0;
     ASSERT_EQ(srt_getsockopt(accepted_sock, 0, SRTO_LOSSMAXTTL, &opt_val, &opt_len), SRT_SUCCESS);
     EXPECT_EQ(opt_val, loss_max_ttl) << "Wrong SRTO_LOSSMAXTTL value on the accepted socket";
-    EXPECT_EQ(opt_len, sizeof opt_len) << "Wrong SRTO_LOSSMAXTTL value length on the accepted socket";
+    EXPECT_EQ(size_t(opt_len), sizeof opt_len) << "Wrong SRTO_LOSSMAXTTL value length on the accepted socket";
 
     SRT_TRACEBSTATS stats;
     EXPECT_EQ(srt_bstats(accepted_sock, &stats, 0), SRT_SUCCESS);
@@ -131,7 +131,7 @@ TEST_F(TestSocketOptions, LossMaxTTL)
 
     ASSERT_EQ(srt_getsockopt(m_listen_sock, 0, SRTO_LOSSMAXTTL, &opt_val, &opt_len), SRT_SUCCESS);
     EXPECT_EQ(opt_val, loss_max_ttl) << "Wrong SRTO_LOSSMAXTTL value on the listener socket";
-    EXPECT_EQ(opt_len, sizeof opt_len) << "Wrong SRTO_LOSSMAXTTL value length on the listener socket";
+    EXPECT_EQ(size_t(opt_len), sizeof opt_len) << "Wrong SRTO_LOSSMAXTTL value length on the listener socket";
 
     ASSERT_NE(srt_close(accepted_sock), SRT_ERROR);
 }
@@ -272,7 +272,7 @@ TEST_F(TestSocketOptions, StreamIDOdd)
     int buffer_len = sizeof buffer;
     EXPECT_EQ(srt_getsockopt(m_caller_sock, 0, SRTO_STREAMID, &buffer, &buffer_len), SRT_SUCCESS);
     EXPECT_EQ(std::string(buffer), sid_odd);
-    EXPECT_EQ(buffer_len, sid_odd.size());
+    EXPECT_EQ(size_t(buffer_len), sid_odd.size());
     EXPECT_EQ(strlen(buffer), sid_odd.size());
 
     StartListener();
@@ -283,7 +283,7 @@ TEST_F(TestSocketOptions, StreamIDOdd)
         buffer[i] = 'a';
     buffer_len = (int)(sizeof buffer);
     EXPECT_EQ(srt_getsockopt(accepted_sock, 0, SRTO_STREAMID, &buffer, &buffer_len), SRT_SUCCESS);
-    EXPECT_EQ(buffer_len, sid_odd.size());
+    EXPECT_EQ(size_t(buffer_len), sid_odd.size());
     EXPECT_EQ(strlen(buffer), sid_odd.size());
 
     ASSERT_NE(srt_close(accepted_sock), SRT_ERROR);
@@ -301,7 +301,7 @@ TEST_F(TestSocketOptions, StreamIDEven)
     int buffer_len = sizeof buffer;
     EXPECT_EQ(srt_getsockopt(m_caller_sock, 0, SRTO_STREAMID, &buffer, &buffer_len), SRT_SUCCESS);
     EXPECT_EQ(std::string(buffer), sid_even);
-    EXPECT_EQ(buffer_len, sid_even.size());
+    EXPECT_EQ(size_t(buffer_len), sid_even.size());
     EXPECT_EQ(strlen(buffer), sid_even.size());
 
     StartListener();
@@ -312,7 +312,7 @@ TEST_F(TestSocketOptions, StreamIDEven)
         buffer[i] = 'a';
     buffer_len = (int)(sizeof buffer);
     EXPECT_EQ(srt_getsockopt(accepted_sock, 0, SRTO_STREAMID, &buffer, &buffer_len), SRT_SUCCESS);
-    EXPECT_EQ(buffer_len, sid_even.size());
+    EXPECT_EQ(size_t(buffer_len), sid_even.size());
     EXPECT_EQ(strlen(buffer), sid_even.size());
 
     ASSERT_NE(srt_close(accepted_sock), SRT_ERROR);
@@ -337,7 +337,7 @@ TEST_F(TestSocketOptions, StreamIDAlmostFull)
     int buffer_len = sizeof buffer;
     EXPECT_EQ(srt_getsockopt(m_caller_sock, 0, SRTO_STREAMID, &buffer, &buffer_len), SRT_SUCCESS);
     EXPECT_EQ(std::string(buffer), sid_amost_full);
-    EXPECT_EQ(buffer_len, sid_amost_full.size());
+    EXPECT_EQ(size_t(buffer_len), sid_amost_full.size());
     EXPECT_EQ(strlen(buffer), sid_amost_full.size());
 
     StartListener();
@@ -348,7 +348,7 @@ TEST_F(TestSocketOptions, StreamIDAlmostFull)
         buffer[i] = 'a';
     buffer_len = (int)(sizeof buffer);
     EXPECT_EQ(srt_getsockopt(accepted_sock, 0, SRTO_STREAMID, &buffer, &buffer_len), SRT_SUCCESS);
-    EXPECT_EQ(buffer_len, sid_amost_full.size());
+    EXPECT_EQ(size_t(buffer_len), sid_amost_full.size());
     EXPECT_EQ(strlen(buffer), sid_amost_full.size());
     EXPECT_EQ(buffer[sid_amost_full.size()-1], 'z');
 
@@ -373,7 +373,7 @@ TEST_F(TestSocketOptions, StreamIDFull)
     int buffer_len = sizeof buffer;
     EXPECT_EQ(srt_getsockopt(m_caller_sock, 0, SRTO_STREAMID, &buffer, &buffer_len), SRT_SUCCESS);
     EXPECT_EQ(std::string(buffer), sid_full);
-    EXPECT_EQ(buffer_len, sid_full.size());
+    EXPECT_EQ(size_t(buffer_len), sid_full.size());
     EXPECT_EQ(strlen(buffer), sid_full.size());
 
     StartListener();
@@ -384,7 +384,7 @@ TEST_F(TestSocketOptions, StreamIDFull)
         buffer[i] = 'a';
     buffer_len = (int)(sizeof buffer);
     EXPECT_EQ(srt_getsockopt(accepted_sock, 0, SRTO_STREAMID, &buffer, &buffer_len), SRT_SUCCESS);
-    EXPECT_EQ(buffer_len, sid_full.size());
+    EXPECT_EQ(size_t(buffer_len), sid_full.size());
     EXPECT_EQ(strlen(buffer), sid_full.size());
     EXPECT_EQ(buffer[sid_full.size()-1], 'z');
